// learnPrisma/src/enums/error-code.type.ts
export enum ERROR_CODE {
  BadRequestError = 3400,
  Unauthorized = 3401,
  Forbidden = 3403,
  DataNotExistError = 3404,
  DataConflict = 3409,
  CreateFailError = 3410,
  DeleteFailError = 3411,
  MissingParamError = 3412,
  UpdateFailError = 3413,
  CreateUserFailError = 3414,
  CreateUserBatchFailError = 3415,
  NameAlreadyExistError = 3416,
  ExceedStorageLimitError = 3417,
  SubDomainAlreadyExistError = 3418,
  PermissionDeniedError = 3420,
  InternalServerError = 3500,
  ExamNotExistError = 7001,
  ReviewNotExistError = 7002,
  ReviewQuestionNotExistError = 7003,
}